회원탈퇴 수정 마이페이지 닉네임뜨도록 수정 배송현황 이제 주문내역에서 따로 버튼 눌러서 페이지이동 배송지 수정

+++ LaClave/src/main/java/com/itwillbs/LaClave/delivery/MyDelivery.java
@@ -31,7 +31,7 @@ public class MyDelivery {
     private Integer memberIdx;
     
     @Column(name = "DELIVERY_STATUS_COMMON_IDX", nullable = false)
-    private Integer deliveryStatusCommonIdx; //공통코드 pk참조idx
+    private Long deliveryStatusCommonIdx; //공통코드 pk참조idx
 	
     @Column(name = "START_DATE", nullable = false)
     private LocalDateTime startDate; //시작 날짜
